requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.1
pip==25.0.1
psycopg2-binary==2.9.3
python-dotenv==1.0.1
setuptools==68.2.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
wheel==0.41.2
djangorestframework==3.15.2
drf-yasg==1.21.8
